+++ CHANGELOG.md
@@ -6,6 +6,12 @@ this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Improved the error response metadata for the universal exception handler in `server.py`.
+- Standardized the 404 error response metadata to include request paths and documentation linked for improved developer experience.
+
 ## [0.7.0] - 2026-04-05
 
 ### Changed