+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improved the error response metadata for the universal exception handler in `server.py`.
 - Standardized the 404 error response metadata to include request paths and documentation linked for improved developer experience.
+- Updated API metadata and reconfigured documentation paths.
 
 ## [0.7.0] - 2026-04-05