+++ CHANGELOG.md
@@ -6,7 +6,7 @@ this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0] - 2026-06-16
 
 ### Added
 - Implemented a real-time database connectivity check in the `/health` endpoint to monitor connection pool status.
@@ -116,6 +116,7 @@ ensure frontend stability.
 - Established `requirements.txt` with core dependencies (FastAPI, Pydantic).
 
 ---
+[0.8.0]: https://github.com/anikethchavare/api.anikethchavare.com/releases/tag/v0.8.0
 [0.7.0]: https://github.com/anikethchavare/api.anikethchavare.com/releases/tag/v0.7.0
 [0.6.1]: https://github.com/anikethchavare/api.anikethchavare.com/releases/tag/v0.6.1
 [0.6.0]: https://github.com/anikethchavare/api.anikethchavare.com/releases/tag/v0.6.0